+++ src/Discore/DiscordActivityFlag.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace Discore
+{
+    [Flags]
+    public enum DiscordActivityFlag
+    {
+        None = 0,
+        Instance = 1 << 0,
+        Join = 1 << 1,
+        Spectate = 1 << 2,
+        JoinRequest = 1 << 3,
+        Sync = 1 << 4,
+        Play = 1 << 5,
+        PartyPrivacyFriends = 1 << 6,
+        PartyPrivacyVoiceChannel = 1 << 7,
+        Embedded = 1 << 8
+    }
+}