+++ src/Discore/Http/DiscordHttpClient.Message.cs
@@ -250,7 +250,7 @@ namespace Discore.Http
 
         /// <summary>
         /// Edits an existing message in a text channel.
-        /// <para>Note: only messages created by the current bot can be editted.</para>
+        /// <para>Note: only messages created by the current bot can be edited.</para>
         /// </summary>
         /// <exception cref="DiscordHttpApiException"></exception>
         public async Task<DiscordMessage> EditMessage(Snowflake channelId, Snowflake messageId, string content)
@@ -283,7 +283,7 @@ namespace Discore.Http
 
         /// <summary>
         /// Edits an existing message in a text channel.
-        /// <para>Note: only messages created by the current bot can be editted.</para>
+        /// <para>Note: only messages created by the current bot can be edited.</para>
         /// </summary>
         /// <exception cref="ArgumentNullException"></exception>
         /// <exception cref="DiscordHttpApiException"></exception>