+++ src/Discore/Http/EditMessageOptions.cs
@@ -14,6 +14,18 @@ namespace Discore.Http
         /// </summary>
         public EmbedOptions? Embed { get; set; }
 
+        /// <summary>
+        /// Gets or sets the allowed mentions for the message.
+        /// </summary>
+        public AllowedMentionsOptions? AllowedMentions { get; set; }
+
+        /// <summary>
+        /// Gets or sets the flags of the message.
+        /// <para/>
+        /// Note: Only <see cref="DiscordMessageFlags.SuppressEmbeds"/> can be set/unset.
+        /// </summary>
+        public DiscordMessageFlags? Flags { get; set; }
+
         public EditMessageOptions() { }
 
         public EditMessageOptions(string content)
@@ -39,6 +51,26 @@ namespace Discore.Http
             return this;
         }
 
+        /// <summary>
+        /// Sets which mentions are allowed for the message.
+        /// </summary>
+        public EditMessageOptions SetAllowedMentions(AllowedMentionsOptions allowedMentions)
+        {
+            AllowedMentions = allowedMentions;
+            return this;
+        }
+
+        /// <summary>
+        /// Sets the flags of the message.
+        /// <para/>
+        /// Note: Only <see cref="DiscordMessageFlags.SuppressEmbeds"/> can be set/unset.
+        /// </summary>
+        public EditMessageOptions SetFlags(DiscordMessageFlags? flags)
+        {
+            Flags = flags;
+            return this;
+        }
+
         internal void Build(Utf8JsonWriter writer)
         {
             writer.WriteStartObject();
@@ -50,6 +82,15 @@ namespace Discore.Http
                 writer.WritePropertyName("embed");
                 Embed.Build(writer);
             }
+
+            if (AllowedMentions != null)
+            {
+                writer.WritePropertyName("allowed_mentions");
+                AllowedMentions.Build(writer);
+            }
+
+            if (Flags != null)
+                writer.WriteNumber("flags", (int)Flags.Value);
 
             writer.WriteEndObject();
         }